Refactoring load and save data
