+++ src/main/java/org/qty/crawler/DataUpdater.java
@@ -12,13 +12,17 @@ import java.io.IOException;
 import java.util.*;
 import java.util.stream.Collectors;
 
+import static org.qty.crawler.IOUtils.loadPreviousTopics;
+
 public class DataUpdater {
 
     public static void main(String[] args) throws IOException {
         Crawler crawler = new Crawler(new DefaultFetch());
+        Storage storage = new DefaultStorage();
 
         List<Topic> topics = crawler.topics();
-        List<Topic> savedTopics = loadPreviousTopics();
+        List<Topic> savedTopics = storage.loadSavedTopics();
+
         if (savedTopics.isEmpty()) {
             savedTopics = topics;
         } else {
@@ -33,15 +37,12 @@ public class DataUpdater {
         });
 
         System.out.println("size: " + savedTopics.size());
-        savedTopics.stream().limit(300).forEach(topic -> {
+        savedTopics.stream().limit(50).forEach(topic -> {
             crawler.update(topic);
             System.out.println(topic);
         });
 
-        Gson gson = new GsonBuilder().setPrettyPrinting().create();
-        Collections.sort(savedTopics, Comparator.comparing(Topic::getUrl));
-        FileUtils.write(new File("data.json"), gson.toJson(savedTopics), "utf-8");
-        FileUtils.write(new File("ui-data.json"), gson.toJson(UIDataModel.convertForUI(savedTopics)), "utf-8");
+        storage.saveTopics(savedTopics);
     }
 
     private static void appendNewTopics(List<Topic> savedTopics, List<Topic> topics) {
@@ -53,16 +54,5 @@ public class DataUpdater {
             }
         });
 
-    }
-
-    public static List<Topic> loadPreviousTopics() throws IOException {
-        if (!new File("data.json").exists()) {
-            return new ArrayList<>();
-        }
-        String data = FileUtils.readFileToString(new File("data.json"), "utf-8");
-        List<Topic> previousTopics = new Gson().fromJson(data, new TypeToken<List<Topic>>() {
-        }.getType());
-
-        return previousTopics;
     }
 }
